complete video 37

+++ backend/src/resolvers/Mutation.js
@@ -45,6 +45,14 @@ const Mutations = {
   async deleteItem(parent, args, ctx, info) {
     const where = { id: args.id }
     const item = await ctx.db.query.item({ where }, `{ id title }`)
+    const ownsItem = item.user.id === ctx.request.userId
+    const hasPermissions = ctx.request.user.permissions.some(permission =>
+      ['ADMIN', 'ITEMDELETE'].includes(permission)
+    )
+
+    if (!ownsItem || !hasPermissions) {
+      throw new Error('You dont have permission')
+    }
 
     return ctx.db.mutation.deleteItem({ where }, info)
   },
